+++ types/Token.d.ts
@@ -1,0 +1,3 @@
+interface Token {
+  refreshToken: string
+}
